refactor: clean up metadata
