+++ packages/react-ui/src/features/pieces/lib/pieces-hook.ts
@@ -30,26 +30,11 @@ type UsePiecesProps = {
   searchQuery?: string;
 };
 
-type PieceStepMetadata = {
+export type StepMetadata = {
   displayName: string;
   logoUrl: string;
   description: string;
-  type: ActionType.PIECE | TriggerType.PIECE;
-  pieceName: string;
 };
-
-type NonPieceStepMetadata = {
-  displayName: string;
-  logoUrl: string;
-  description: string;
-  type:
-    | ActionType.BRANCH
-    | ActionType.CODE
-    | ActionType.LOOP_ON_ITEMS
-    | TriggerType.EMPTY;
-};
-
-export type StepMetadata = NonPieceStepMetadata | PieceStepMetadata;
 
 export const piecesHooks = {
   usePiece: ({ name, version }: UsePieceProps) => {
@@ -79,7 +64,7 @@ export const piecesHooks = {
     const pieceVersion = step.settings?.pieceVersion;
     const query = useQuery<StepMetadata, Error>({
       queryKey: ['piece', type, pieceName, pieceVersion],
-      queryFn: () => getStepMetadata(step),
+      queryFn: () => piecesApi.getMetadata(step),
       staleTime: Infinity,
       enabled,
     });
@@ -109,55 +94,7 @@ function stepMetadataQueryBuilder(step: Step) {
   const pieceVersion = isPieceStep ? step.settings.pieceVersion : undefined;
   return {
     queryKey: ['piece', step.type, pieceName, pieceVersion],
-    queryFn: () => getStepMetadata(step),
+    queryFn: () => piecesApi.getMetadata(step),
     staleTime: Infinity,
   };
-}
-
-async function getStepMetadata(step: Step): Promise<StepMetadata> {
-  switch (step.type) {
-    case ActionType.BRANCH:
-      return {
-        displayName: 'Branch',
-        logoUrl: 'https://cdn.activepieces.com/pieces/branch.svg',
-        description: 'Branch',
-        type: step.type,
-      };
-    case ActionType.CODE:
-      return {
-        displayName: 'Code',
-        logoUrl: 'https://cdn.activepieces.com/pieces/code.svg',
-        description: 'Powerful nodejs & typescript code with npm',
-        type: ActionType.CODE,
-      };
-    case ActionType.LOOP_ON_ITEMS:
-      return {
-        displayName: 'Loop on Items',
-        logoUrl: 'https://cdn.activepieces.com/pieces/loop.svg',
-        description: 'Iterate over a list of items',
-        type: ActionType.LOOP_ON_ITEMS,
-      };
-    case TriggerType.EMPTY: {
-      return {
-        displayName: 'Empty Trigger',
-        logoUrl: 'https://cdn.activepieces.com/pieces/empty-trigger.svg',
-        description: 'Empty Trigger',
-        type: TriggerType.EMPTY,
-      };
-    }
-    case ActionType.PIECE:
-    case TriggerType.PIECE: {
-      const piece = await piecesApi.get({
-        name: step.settings.pieceName,
-        version: step.settings.pieceVersion,
-      });
-      return {
-        displayName: piece.displayName,
-        logoUrl: piece.logoUrl,
-        description: piece.description,
-        type: step.type,
-        pieceName: step.settings.pieceName,
-      };
-    }
-  }
 }
